rename files for solar
rename files for parabolic trough
