--- labothappy/component/solar/parabolic_through/design/parabolic_plant_design.py
@@ -1,11 +1,0 @@
-import __init__
-import numpy as np
-import labothappy
-import labothappy.component
-from labothappy.geometries.solar.parabolic_through_geometry import PT_Collector_Geom
-# from library.component.sizing.solar.parabolic_through.
-
-PT_geom = PT_Collector_Geom()
-PT_geom.set_parameters("Soponova_MicroCSP")
-
-
