+++ labothappy/geometries/solar/parabolictrough_geometry.py
